fixed the file arrangement
